GH-26 Add basic HTB commands for now
