+++ bot/src/commands/mod.rs
@@ -1,0 +1,2 @@
+pub mod ctf;
+pub mod htb;
